fix bug on 3.0+, remove spurious pc-relative
Oh, this was nasty. Apparently, OS 1 - 2 zero the first 24 bits of d5. This is not true in 3.0, so a move.b #$99 means the first 24 bits hold garbage. This is ok, except we then move that bit to d6, WHICH IS NOT CLEARED, and apply bitshifts to it. The end result is that bits 8 and 7 get values from previous data in d6.

Most economical way of fixing this was by simply doing a move.l instead.